@@ -11,14 +11,19 @@
         {
             "type": "vcs",
             "url": "https://github.com/adamwathan/laravel-preset"
+        },
+        {
+            "type": "vcs",
+            "url": "https://github.com/AdrianBav/traffic"
         }
     ],
     "require": {
         "php": "^7.1.3",
         "fideloper/proxy": "^4.0",
-        "laravel/framework": "5.7.*",
+        "laravel/framework": "5.8.*",
         "laravel/tinker": "^1.0",
-        "writingink/wink": "^0.0.12"
+        "writingink/wink": "^0.0.12",
+        "adrianbav/traffic": "dev-master"
     },
     "require-dev": {
         "adamwathan/laravel-preset": "dev-master",
@@ -26,8 +31,8 @@
         "filp/whoops": "^2.0",
         "fzaninotto/faker": "^1.4",
         "mockery/mockery": "^1.0",
-        "nunomaduro/collision": "^2.0",
-        "phpunit/phpunit": "^7.0"
+        "nunomaduro/collision": "^3.0",
+        "phpunit/phpunit": "^7.5"
     },
     "config": {
         "optimize-autoloader": true,
